usb-loader: add method to force FULL page mode
If the DF at45db641e has been shipped in binary mode it will not work
fine with our partition system.

> mtd: partition "hwinfo" doesn't end on an erase block -- force read-
only

In this sense we need to ensure that we are in full page mode and not
binary mode by executing:

     mch_flasher -m df -c

If the WRS is already in full mode nothing will be done.