build/download: make wget timeout sooner and avoid retrying
The upstream address for lsof is not working. While the download script
tries buildroot.net if upstream fails, getting lsof got stuck for
no connection to the server.

With this commit it times out after 15 seconds to then proceed using
the buildroot address. I expect the problem to affect other upstream
files over time, so the commit doesn't fix the link but the tool.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>