rootfs: fix lack of reboot after update
The problem is that the "reboot" command sends the signal to the init. However,
it looks like init queues all incoming signals. They're handled after all
"sysinit" commands from the inittab are finished.
reboot -f skip sending the signal and triggers the reboot immediately.

This problem pop up after update of a buildroot to version 2016.02.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>