[Feature: 1363] rootfs: add a support to execute custom script at startup
This will allow users to have special configuration (that is not possible to
achieve by just using dot-config) without a need to manually handle custom
script.

The script can be used from /wr/bin/custom_boot_script.sh or downloaded during
the boot. It is not possible to execute a script which was downloaded during
the previous boot. But a script can be written in such way to copy itself to
/wr/bin/custom_boot_script.sh.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>