[Bug: 1219] rootfs/www: fix problem with uploading the firmware
PHP code of the web interface was updating values of "upload_max_filesize"
and "post_max_size" in the /etc/php.ini. Besides it is extremely bad idea to
do that, such changes to take place require restart of lighttpd, which was
never performed. Luckily after the update of /etc/php.ini it was also copied
to flash at the location /usr/etc/php.ini. Because of this update via
webinterface worked only when:
--firmware.php was opened via web browser
--restart of a switch was performed

The fix contains a patch to set these values in the php.ini and remove
the update of php.ini by php code.

Set maximum upload limit size to 30M in php.ini.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>