userspace/tools: clear VID=0 when set up VLANs from dot-config
By default, on the startup the VLAN with VID0 is created with the mask
0xffffffff. It is used to pass all the traffic between ports. However, when
VLANs configuration is read from the dot-config now it also remove this VLAN.
We cannot relay on how the system was configured before.
If such configuration is needed it can be configured in the
CONFIG_VLANS_VLAN0000 config option.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>