kernel: fix headers
Many mach headers has been moved out of scope. Of course they are still
accessible by in an ugly way. Here I'm following the ugly way in order
to get something working.

What to do with this headers, perhaps, needs a discussion. I personally
think that we need our local headers with hardware definitions. The
reason is that we saw that in future kernel version some of these
headers complitely disappear. The hardware definitions never change so
we can safely copy them and remove a dependency with an external project.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>