userspace:init: load g_serial new dependencies
The USB gadget framework has changed from Linux v2.6.39. As far as I
understood it's now more modular, so it's not anymore enough to load
just the g_serial.ko driver. Now we have to load also it's
dependencies.

I'm not an expert about USB gadget, this patch is the result of my
understanding in debugging the loading failure of the g_serial.ko
module. Perhaps it will be necessary some adjustments to make it
work the way we want. But for the time begin I will leave it like
this until is clear what we have to adjust.

Note that these dependecies depends also on the configfs file-system.
I had problems to get this work because apparently by using a different
corss-compiler some symbols where removed from the kernel, so our
modules were not able to load because of "undefined symbols".
Using the toolchain compiled within this project solved this issue.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>