userspace:init: remove everything from /dev/shm at boot
We want '/dev/shm' to be empty at boot time. It may happen that
while working with rootfs over NFS this directory is not empty.

With this patch we remove all files from this directory because
they should not be there at all.

What if I put something there on purpose? When using tmpfs,
files in '/dev/shm' will be automatically removed on shutdown.
We are keeping the same behavior. Put your own development files
in '/dev/shm' after the execution of this script by copying them
from another location.

Again, if we use rootfs over NFS and if tmpfs does not work then
the system will use the NFS and the system should continue to work.
In this case the performances will be much worst

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>