kernel:*: get IRQ numbers from the VIC domain
Actually, this is not what I want. I would like to just use an IRQ
resource with the correct IRQ number. What I want is to hide the IRQ
domain knowledge elsewhere. For instance, in the wrs_device module
that register all our devices. But, since I have to update the kernel
with the least number of changes I will leave this for the future.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>