userspace:tools: fix linux headers use
The Linux kernel 3.16.37 does not export, or does not have, anymore some
hardware headers. Here I copy all the values we need.

The real fix should be to create a set of WRS headers with all the symbols
we need so that we do not depend anymore on other projects.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>