userspace:tools: including kernel header is dangerous, remove then
It happens that headers have different sources: buildroot and our own kernel.
For the standard path we have headers from buildroot which include also
some kernel header. For the kernel, we have direct access to those headers.

this means that we mya have conflict in our user-space programs. This patch
does not solve the problem but it just reduce its effects.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>