[BUG: 1552] proto-ext-whiterabbit: add a pre-master state to wr fsm table
When two masters are connected to the same link one of them tries to enter
a state pre-master. However, this state is not compiled in the ppsi, so one
node stuck there forever. It is necessary to restart a node or re-establish
a link.

BUG introduced by a PPSI's commit:
2996dd7b compliance, 9.2.6.10: properly switch to MASTER or PRE_MASTER

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>