[BUG: 1590] userspace/snmpd: fix fail to start when using non-wr SFPs
When many (10 or more) non-wr compatible SFPs were used then could happen that
snmpd was not able to start (sometimes even crashing).
The problem appeared only when the logs from snmpd were not redirected,
so it too long to print the logs to the slow serial console.
Before logs were printed, the content of HAL's shmem changed and this caused
to re-read HAL's shmem and print logs again and again...

The fix is to print logs after the shmem is read.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>